ENH: Consistently use strongly typed enums
CHANGED:
  - AreaType to AreaEnum enum class
  - ConnectivityEnumType to ConnectivityEnum enum class
  - EdgeStatusType to EdgeStatusEnum enum class
  - GradientType to GradientEnum enum class
  - GrowthStrategyType to GrowthStrategyEnum enum class
  - GrowthStrategyType to GrowthStrategyEnum enum object
  - HashFunction to HashFunctionEnum enum class
  - Include Legacy_Remove parameters for backwards
  - InterpolationMethodType to InterpolationMethodEnum enum class
  - InterpolationModeType into InterpolationModeEnum enum class
  - LabelType to LabelEnum enum class
  - LeafIdentifier to LeafIdentifierEnum enum class
  - LineSearchMethod to LineSearchMethodEnum
  - LogLevel to LogLevelEnum enum class
  - MetricCategoryType to MetricCategoryEnum enum class
  - MetricSamplingStrategyType to MetricSamplingStrategyEnum
  - NoiseModelType to NoiseModelEnum enum object
  - OctreePlaneType to OctreePlaneEnum enum class
  - OctreePlaneType to OctreePlaneNode enum object
  - RegionType to RegionEnum enum class
  - SIDE to SIDEEnum enum class
  - SigmaStepMethodType to SigmanStepMethodEnum enum class
  - Sign to SignEnum enum class
  - StopConditionEnum to scoped enum class
  - TERMINATION_CODE to TERMINATION_CODE_ENUM enum class
  - TargetConditionType to TargetConditionEnum enum class
  - TemporalUnitType to TemporalUnitEnum enum class
  - TimeStampFormatType to TimeStampFormatEnum enum class
  - TopologyCheckType to TopologyCheckEnum enum class
  - TransformDirectionType to TransformDirectionEnum enum class
  - TraversalDirection to TraversalDirectionEnum enum class
  - WeightType to WeightEnum enum object
  - ConnectedThresholdImageFilterEnums/ -> ConnectivityEnum
  - QuadEdgeMeshEulerOperatorFlipEdgeFunctionEnums/EdgeStatusType -> EdgeStatusEnum
  - ObjectStoreEnums/GrowthStrategy -> GrowthStrategyEnum
  - HashImageFilterEnums/HashFunction -> HashFunctionEnum
  - CommonEnums
    + IOPixel -> IOPixelEnum
    + IOComponent -> IOComponentEnum
    + IOFile -> IOFileEnum
    + IOFileMode -> IOFileModeEnum
    + IOByteOrder -> IOByteOrderEnum
    + CellGeometry -> CellGeometryEnum
  - MeshEnums
    + MeshClassCellsAllocationMethod -> MeshClassCellsAllocationMethod
  - ContourSpatialObjectEnums/InterpolationMethod -> InterpolationMethodEnum
  - GaussianDerivateOperatorEnums/InterpolationMode -> InterpolationModeEnum
  - OctreeNodeEnums/LeafIdentifier -> LeafIdentifierEnum
  - LBFGS2Optimizerv4/LineSearchMethod -> LineSearchMethodEnum
  - DCMTKImageIOEnums/LogLevel -> LogLevelEnum
  - OctreeBaseEnums/Octree -> OctreeBaseEnums
  - RegionEnums/RegionEnum -> RegionEnum
  - FastMarchingTraitsEnum/TopologyCheck -> TopologyCheckEnum
  - LoggerBaseEnums/(PriorityLevel & TimeStampFormat) -> PriorityLevelEnum & TimeStampFormatEnum
  - TemporalDataObjectEnums/TemporalUnit -> TemporalUnitEnum
  - ComplexToComplexFFTImageFilterEnums/TransformDirection -> TransformDirectionEnum
  - ScalarToRGBColormapImageFilterEnums/RGBColormapFilter -> RGBColormapFilterEnum
  - FRPROptimizerEnums/Optimization -> Optimization
  - TreeIteratorBaseEnums/TreeIteratorBaseNode -> NodeType (NO TEST NEED NEW TEST)
  - FloatingPointExceptionsEnums/ExceptionAction -> ExceptionActionEnum
  - VideoIOFactoryEnums/IOMode -> IOModeEnum
  - FastMarchingreachedTargetNoesStoppingCriterionEnums/TargetCondition -> TargetConditionEnum
  - ObjectFactoryBaseEnums/InsertionPosition-> InsertionPositionEnums
  - HistogramToRunLengthFeaturesFilterEnums/RunLengthFeature -> RunLengthFeatureEnum
  - MultiThreaderBaseEnums/ThreadExitCode -> ThreadExitCodeEnum
  - MultiThreaderBaseEnums/Threader -> ThreadEnum
  - DTITibeSpatialObjectPointenums/DTITubeSpatialObjectPointField -> DTITubeSpatialObjectPointFieldEnum
  - ConvolutionImageFilterBaseEnums/ConvolutionImageFilterOutput -> ConvolutionImageFilterOutputEnum
  - ExtractSliceImageFilterEnums/TestExtractSliceImageFilterCollapseStrategy -> TestExtractSliceImageFilterCollapseStrategyEnum
  - PatchBasedDenoisingBaseImageFilterEnums/
    + NoiseModel -> NoiseModelEnum
    + ComponentSpace -> ComponentSpaceEnum
    + FilterState -> FilterStateEnum
  - DiffusionTensor3DReconstructionImageFilterEnums/GradientImageFormat -> GradientImageFormatEnum
  - GDCMImageIOEnums/Compression -> CompressionEnums
  - FastMarchingImageFilterEnums/Label -> LabelEnums
  - VideoIOBaseEnums/ReadFrom -> ReadFromEnum
  - ConnectThresholdImageFilterEnums/Connectivity -> ConnectivityEnum
  - DeformableSimplexMesh3DGradientConstraintForceFilter/SIDE -> SIDEEnum
  - ImageRegistrationMethodv4Enums/MetricSamplingStrategy -> MetricSamplingStrategyEnum
  - TransformBaseTemplateEnums/TransformCategory -> TransformCategoryEnums
  - MergeLabelMapFilterEnums/ChoiceMethod -> ChoiceMethodEnum
  - LaplacianDeformationQuadEdgeMeshFilterEnums/Area -> AreaEnum
  - IOCommonEnums/AtomicPixel -> AtomicPixelEnum
  - ESMDemonsRegistrationFunctionEnums/Gradient -> GradientEnum
  - RecursiveGaussianImageFilterEnums/GaussianOrder -> GaussianOrderEnum
  - SPSAOptimizerEnums/StopConditionSPSAOptimizer -> StopConditionSPSAOptimizerEnum
  - SymmetricEigenAnalysisEnums/EigenValueOrder -> EigenValueOrderEnum
  - ObjectToObjectMetricBaseTemplateEnums
    + GradientSource -> GradientSourceEnums
    + MetricCategory -> MetricCategoryEnums
  - ObjectToObjectOptimizerBaseTemplateEnums/StopConditionObjectToObjectOptimizer -> StopConditionObjectToObjectOptimizerEnum
  - RegistrationParameterScalesEstimatorEnums/SamplingStrategy -> SamplingStrategyEnum
  - HistogramToTextureFeaturesFilterEnums/TextureFeature -> TextureFeatureEnum
  - RayCasterHelperEnums/TraversalDirection -> TraversalDirectionEnum
  - CallRecordEnums/
    + RecordType -> RecoredEnum
    + MethodType -> MethodEnum
  - JPEG2000ImageIOInternalEnums/
    + DecodingFormat -> DecodingFormatEnum
    + DFMFormat -> DFMFormaType
  - FEMRegistrationFilterEnums/Sign -> SignEnum

CREATION:
  - new enum class ExceptionEnum
  - ExceptionActionEnum enum object
  - MetricSamplingStrategyEnum enum class (scoped)
  - Replace old static_cast with new itkExposeEnumValue to faciliate
    future refactorings

NEW NAMES: These enums have been consolidated so only one
           definition exists across the toolkit
  - Image/Mesh/Transform versions of FileEnum|FileModeEnum|ByteOrderEnum
    are now named IO(FileEnum|FileModeEnum|ByteOrderEnum) in itk namespace
  - CellGeometryEnum is moved to itk namespace to remove exact duplicate
    definitions in multiple modules (IO/Quadedgemesh/...)

NEW FILE:
EnumPrintFunction.py - creates a basic template of an ITKv5 enum class

Change some duplicate enum class names that have different values so
that they are uniquely named for their intended purpose.

Co-Authored by: Mathew J. Seng <mathewseng@gmail.com>
Co-Authored by: Matt McCormick <matt.mccormick@kitware.com>
Co-Authored by: Hans Johnson <hans-johnson@uiowa.edu>

+++ Modules/Core/Mesh/test/itkCellInterfaceTest.cxx
@@ -48,7 +48,7 @@ TestCellInterface(std::string name, TCell * aCell)
   CellAutoPointer cell(aCell, true);
 
   std::cout << "-------- " << name << " (" << aCell->GetNameOfClass() << ")" << std::endl;
-  std::cout << "    Type: " << cell->GetType() << std::endl;
+  std::cout << "    Type: " << static_cast<int>(cell->GetType()) << std::endl;
   std::cout << "    Dimension: " << cell->GetDimension() << std::endl;
   std::cout << "    NumberOfPoints: " << cell->GetNumberOfPoints() << std::endl;
   std::cout << "    NumberOfBoundaryFeatures:" << std::endl;
